+++ Shooter/Shooter/Shooter/PatternClasses/EnemiesConcreteFactory.cs
@@ -8,7 +8,7 @@ namespace Shooter.PatternClasses
 {
     public class EnemiesConcreteFactory : EnemiesFactory
     {
-        public override Enemy CreateEnemy(IPathFinding pathFinder, EnemyType type, IWeapon weapon, IPlayer player, int lifePoints, Vector2 position, Texture2D texture)
+        public override Enemy CreateEnemy(IPathFinding pathFinder, EnemyType type, IWeapon weapon, IPlayer player, double lifePoints, Vector2 position, Texture2D texture)
         {
             switch (type)
             {
